auth/requirements/base.txt
# Django
Django==3.1.1

# Rest Framework
djangorestframework==3.11.1
django-filter==2.3.0

# Routers
drf-nested-routers==0.91

# Run Server
gunicorn[gevent]==20.0.4
whitenoise==5.2.0

# CORS
django-cors-headers==3.5.0

#JWT
djangorestframework-simplejwt==5.0.0